Fix 64-bit compile problems - %d doesn't work with size_t, so instead use %ld and casts to long.

+++ src/escputil/escputil.c
@@ -1,5 +1,5 @@
 /*
- * "$Id: escputil.c,v 1.96 2008/07/05 01:43:26 rlk Exp $"
+ * "$Id: escputil.c,v 1.97 2008/10/29 00:10:42 easysw Exp $"
  *
  *   Printer maintenance utility for EPSON Stylus (R) printers
  *
@@ -1252,9 +1252,9 @@ do_old_status(status_cmd_t cmd, const char *buf, const stp_printer_t *printer)
 		{
 		  color_list = stp_string_list_create_copy(desc.bounds.str);
 		  STP_DEBUG(fprintf(stderr,
-				    "Using color list from driver (%d %d)\n",
-				    stp_string_list_count(desc.bounds.str),
-				    stp_string_list_count(color_list)));
+				    "Using color list from driver (%ld %ld)\n",
+				    (long)stp_string_list_count(desc.bounds.str),
+				    (long)stp_string_list_count(color_list)));
 		  stp_parameter_description_destroy(&desc);
 		}
 	    }
@@ -1520,9 +1520,9 @@ do_extended_ink_info(int extended_output)
       if (printer)
 	{
 	  color_list = stp_string_list_create_copy(desc.bounds.str);
-	  STP_DEBUG(fprintf(stderr, "Using color list from driver (%d %d)\n",
-			    stp_string_list_count(desc.bounds.str),
-			    stp_string_list_count(color_list)));
+	  STP_DEBUG(fprintf(stderr, "Using color list from driver (%ld %ld)\n",
+			    (long)stp_string_list_count(desc.bounds.str),
+			    (long)stp_string_list_count(color_list)));
 	  stp_parameter_description_destroy(&desc);
 	}
       else
